--- a/.github/workflows/perf-v2.js.yml
+++ b/.github/workflows/perf-v2.js.yml
@@ -44,7 +44,7 @@ jobs:
           # Workflow will fail when an alert happens
           fail-on-alert: true
           github-token: ${{ secrets.PERSONAL_GITHUB_TOKEN }}
-          benchmark-data-dir-path: perf
+          benchmark-data-dir-path: perf-v2
           auto-push: true
           alert-threshold: '130%'
           comment-on-alert: true