--- a/perf/format.ts
+++ b/perf/format.ts
@@ -19,8 +19,11 @@ export function formatAsBenchmarkJS(results: BenchmarkResult): string {
   const value = results.byteSize
     ? formatToMBPerSecond(results.byteSize, medianMs)
     : `${((1 / medianMs) * 1000).toFixed(2)} ops/sec`;
-  const variance = results.runTimesStatistics.variance.toFixed(1);
-  return `${name} x ${value}±${variance}% (${runs.length} runs sampled)`;
+  return `${name} x ${value}${formatVariance(stats.variance)} (${runs.length} runs sampled)`;
+}
+
+export function formatVariance(variance: number): string {
+  return `±${variance.toFixed(1)}%`
 }
 
 function formatToMBPerSecond(size: number, timeMS: number): string {