--- a/perf/format.ts
+++ b/perf/format.ts
@@ -19,7 +19,7 @@ export function formatAsBenchmarkJS(results: BenchmarkResult): string {
   const value = results.byteSize
     ? formatToMBPerSecond(results.byteSize, medianMs)
     : `${((1 / medianMs) * 1000).toFixed(2)} ops/sec`;
-  return `${name} x ${value}${formatVariance(stats.variance)} (${
+  return `${name} x ${value} ${formatVariance(stats.variance)} (${
     runs.length
   } runs sampled)`;
 }