--- a/perf/perf.ts
+++ b/perf/perf.ts
@@ -3,7 +3,7 @@ import {benchmarkIDBRead, benchmarkIDBWrite} from './idb';
 import {benchmarks as lockBenchmarks} from './lock';
 import {benchmarks as hashBenchmarks} from './hash';
 import {benchmarks as storageBenchmarks} from './storage';
-import {formatAsBenchmarkJS, formatAsReplicache} from './format';
+import {formatAsBenchmarkJS, formatAsReplicache, formatVariance} from './format';
 import type {RandomDataType} from './data';
 
 export type Benchmark = {
@@ -221,7 +221,7 @@ function createGithubActionBenchmarkJsonEntry(result: BenchmarkResult): Entry {
     name: result.name,
     unit: 'median ms',
     value: result.runTimesStatistics.medianMs,
-    range: result.runTimesStatistics.variance.toFixed(2).toString(),
+    range: formatVariance(result.runTimesStatistics.variance),
     extra: formatAsReplicache(result),
   };
 }