--- a/perf/perf.ts
+++ b/perf/perf.ts
@@ -3,7 +3,11 @@ import {benchmarkIDBRead, benchmarkIDBWrite} from './idb';
 import {benchmarks as lockBenchmarks} from './lock';
 import {benchmarks as hashBenchmarks} from './hash';
 import {benchmarks as storageBenchmarks} from './storage';
-import {formatAsBenchmarkJS, formatAsReplicache, formatVariance} from './format';
+import {
+  formatAsBenchmarkJS,
+  formatAsReplicache,
+  formatVariance,
+} from './format';
 import type {RandomDataType} from './data';
 
 export type Benchmark = {