This does not get rid of the existing perf dashboard at https://rocicorp.github.io/replicache/perf/ and the alerts based on that, but rather adds a new one at https://rocicorp.github.io/replicache/perf-v2/ with corresponding alerts.  This ensures we have continuity of our benchmark data.  Once sufficient history is built up on the new dashbaord, we can clean up the old one (which will also allow some simplification/cleanup of the related code).  In the meantime we will get double alerts.

This utilizes the `customSmallerIsBetter` tool option of github-action-benchmark (see https://github.com/benchmark-action/github-action-benchmark#examples).

We have been using the `benchmarkjs` tool option, and emulating the output of that library https://github.com/bestiejs/benchmark.js, though we are not actually using the library.  Rather we are using our own custom benchmark implementation.

Example JSON output:

```
Gregorys-MacBook-Pro:replicache greg$ node perf/runner.js --format=json --run "scan"
[
  {
    "name": "scan 1024x1000",
    "unit": "median ms",
    "value": 52.5,
    "range": "8.70",
    "extra": "scan 1024x1000 50/75/90/95%=52.50/54.40/59.30/59.30 ms avg=51.34 ms (10 runs sampled)"
  },
  {
    "name": "scan 1024x5000",
    "unit": "median ms",
    "value": 236.5,
    "range": "27.50",
    "extra": "scan 1024x5000 50/75/90/95%=236.50/246.40/264.00/264.00 ms avg=242.74 ms (5 runs sampled)"
  },
  {
    "name": "[MemStore] scan 1024x1000",
    "unit": "median ms",
    "value": 1.7000000178813934,
    "range": "4.10",
    "extra": "[MemStore] scan 1024x1000 50/75/90/95%=1.70/2.00/2.20/2.30 ms avg=1.86 ms (269 runs sampled)"
  },
  {
    "name": "[MemStore] scan 1024x5000",
    "unit": "median ms",
    "value": 10.300000011920929,
    "range": "7.10",
    "extra": "[MemStore] scan 1024x5000 50/75/90/95%=10.30/10.60/11.20/11.30 ms avg=10.53 ms (48 runs sampled)"
  }
]
```