@grgbkr 2021-11-02
@arv I think as I have perf.js.yml and perf-v2.js.yml configured, the two workflows should not interfere with each other, but I'm not 100% sure.  I don't know how to test this without landing the pr.  Do you know of a way to test?