@arv 2021-11-02
> Do you know of a way to test?

Two ways:

1. You can install a local github action runner (I have never done this but @whiten did it at one point)
2. You can temporary change the yml file to run on every PR (and skip the commit to the branch)