@aboodman 2021-11-02
I’m ok just comparing to a screencap for awhile if that helps move forward
faster.

On Tue, Nov 2, 2021 at 7:19 AM Erik Arvidsson ***@***.***>
wrote:

> Do you know of a way to test?
>
> Two ways:
>
>    1. You can install a local github action runner (I have never done
>    this but @whiten <https://github.com/whiten> did it at one point)
>    2. You can temporary change the yml file to run on every PR (and skip
>    the commit to the branch)
>
> —
> You are receiving this because your review was requested.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/rocicorp/replicache/pull/654#issuecomment-957963450>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAATUBCV6SBOEGT7Q7B7ZQTUKATUFANCNFSM5HG5BAEA>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>
-- 
a (phone)