@aboodman 2021-11-02
Removing myself from review - I think erik can handle this. If it's even a medium pita to maintain continuity I'm fine losing the old data in the interest of velocity.